exc/requirements.txt
alembic
cryptography
fastapi
pymysql
python-dotenv
SQLAlchemy
uvicorn